requirements.txt
pyconll
pytest
pyyaml
toolz